+++ src/model/DeviceState.h
@@ -25,7 +25,11 @@ struct MelodyNote {
 #define SCREEN_RADEC      2
 #define SCREEN_ALTAZ      3
 #define SCREEN_BATTERY    4
-#define SCREEN_MESSAGE    5
+#define SCREEN_MESSAGE     5
+#define SCREEN_SYSINFO_1   6   // secondary pages; reachable only via BtnB short press from BATTERY
+#define SCREEN_SYSINFO_2   7
+#define SCREEN_SYSINFO_3   8
+#define SCREEN_SYSINFO_LAST SCREEN_SYSINFO_3
 
 // Button bitmask constants
 #define BTN_MASK_M5  0x01